♻️ボス固有状態異常の分離とファイル構成の改善
- battle-effects.tsからボス固有の状態異常を分離
- 各ボス専用の状態異常ファイルを新規作成:
  - scorpion-carrier-effects.ts (麻酔、サソリ毒、脱力剤)
  - mikan-dragon-effects.ts (脱力)
  - sea-kraken-effects.ts (視界阻害)
  - aqua-serpent-effects.ts (水浸し、目眩)
  - clean-master-effects.ts (泡まみれ、回転中、蒸し暑い)
  - dark-ghost-effects.ts (恐怖、忘却)
  - underground-worm-effects.ts (石化)
- dream-demon-effects.tsに夢操作状態異常を移動
- battle-effects.tsは汎用戦闘状態異常のみに整理
- index.tsで新しい状態異常ファイルを自動登録
