+++ src/game/systems/status-effects/battle-effects.ts
@@ -157,213 +157,5 @@ export const battleEffectsConfigs: StatusEffectConfig[] = [
         modifiers: {
             canAct: false
         }
-    },
-    {
-        type: StatusEffectType.DreamControl,
-        name: '夢操作',
-        description: '夢の世界に閉じ込められ、完全に支配されている',
-        duration: -1, // Permanent while sleeping
-        category: 'debuff',
-        isDebuff: true,
-        modifiers: {
-            canAct: false,
-            canUseSkills: false
-        }
-    },
-
-    // Scorpion Carrier status effects
-    {
-        type: StatusEffectType.Anesthesia,
-        name: '麻酔',
-        description: '麻酔により行動不能',
-        duration: 3,
-        category: 'debuff',
-        isDebuff: true,
-        modifiers: {
-            canAct: false
-        }
-    },
-    {
-        type: StatusEffectType.ScorpionPoison,
-        name: 'サソリ毒',
-        description: '毎ターン最大HPの1/10のダメージを受ける',
-        duration: 3,
-        category: 'debuff',
-        isDebuff: true,
-        stackable: true,
-        onTick: (target: Actor, _effect: StatusEffect) => {
-            const damage = Math.floor(target.maxHp / 10);
-            target.takeDamage(damage);
-        },
-        messages: {
-            onApplyPlayer: '{name}は強力なサソリ毒に冒された！',
-            onApplyBoss: '{name}は強力なサソリ毒に冒された！',
-            onTickPlayer: '{name}はサソリ毒により{damage}のダメージを受けた！',
-            onTickBoss: '{name}はサソリ毒により{damage}のダメージを受けた！',
-            onRemovePlayer: '{name}の体からサソリ毒が抜けた',
-            onRemoveBoss: '{name}の体からサソリ毒が抜けた'
-        }
-    },
-    {
-        type: StatusEffectType.Weakening,
-        name: '脱力剤',
-        description: '体力が奪われ、受けるダメージが1.5倍になる',
-        duration: 3,
-        category: 'debuff',
-        isDebuff: true,
-        modifiers: {
-            damageReceived: 1.5
-        }
-    },
-
-    // Mikan Dragon status effects
-    {
-        type: StatusEffectType.Lethargy,
-        name: '脱力',
-        description: '体に力が入らず、攻撃力が大幅に低下',
-        duration: 3,
-        category: 'debuff',
-        isDebuff: true,
-        modifiers: {
-            attackPower: 0.3
-        }
-    },
-
-    // Sea Kraken status effects
-    {
-        type: StatusEffectType.VisionImpairment,
-        name: '視界阻害',
-        description: 'イカスミで視界が阻害され、攻撃の命中率が低下',
-        duration: 3,
-        category: 'debuff',
-        isDebuff: true,
-        modifiers: {
-            accuracy: 0.7 // Reduces accuracy to 70% of its original value
-        }
-    },
-
-    // Aqua Serpent status effects
-    {
-        type: StatusEffectType.WaterSoaked,
-        name: '水浸し',
-        description: '水で濡れて防御力が低下',
-        duration: 3,
-        category: 'debuff',
-        isDebuff: true,
-        modifiers: {
-            damageReceived: 1.3 // Increases damage received by 30%
-        }
-    },
-    {
-        type: StatusEffectType.Dizzy,
-        name: '目眩',
-        description: 'ふらつきで攻撃の命中率が低下',
-        duration: 2,
-        category: 'debuff',
-        isDebuff: true,
-        modifiers: {
-            accuracy: 0.8 // Reduces accuracy to 80% of its original value
-        }
-    },
-
-    // Clean Master status effects
-    {
-        type: StatusEffectType.Soapy,
-        name: '泡まみれ',
-        description: '石鹸の泡で滑りやすくなり、命中率が低下',
-        duration: 3,
-        category: 'debuff',
-        isDebuff: true,
-        modifiers: {
-            accuracy: 0.7, // Reduces accuracy to 70%
-            struggleRate: 0.8 // Slightly harder to struggle
-        }
-    },
-    {
-        type: StatusEffectType.Spinning,
-        name: '回転中',
-        description: '遠心分離で回転し、命中率が大幅低下',
-        duration: 2,
-        category: 'debuff',
-        isDebuff: true,
-        modifiers: {
-            accuracy: 0.5, // Reduces accuracy to 50%
-            canAct: false // Cannot act during first turn
-        }
-    },
-    {
-        type: StatusEffectType.Steamy,
-        name: '蒸し暑い',
-        description: '温風で蒸し暑くなり、毎ターン軽微ダメージ',
-        duration: 3,
-        category: 'debuff',
-        isDebuff: true,
-        stackable: true,
-        onTick: (target: Actor, _effect: StatusEffect) => {
-            target.takeDamage(4);
-        },
-        messages: {
-            onApplyPlayer: '{name}は蒸し暑さに包まれた！',
-            onApplyBoss: '{name}は蒸し暑さに包まれた！',
-            onTickPlayer: '{name}は熱風により{damage}のダメージを受けた！',
-            onTickBoss: '{name}は熱風により{damage}のダメージを受けた！',
-            onRemovePlayer: '{name}を包む蒸し暑さが収まった',
-            onRemoveBoss: '{name}を包む蒸し暑さが収まった'
-        }
-    },
-
-    // Dark Ghost mental effects
-    {
-        type: StatusEffectType.Fear,
-        name: '恐怖',
-        description: '恐怖により受けるダメージが1.5倍になる',
-        duration: 3,
-        category: 'debuff',
-        isDebuff: true,
-        modifiers: {
-            damageReceived: 1.5
-        },
-        messages: {
-            onApplyPlayer: '{name}は恐怖に支配された！',
-            onApplyBoss: '{name}は恐怖に支配された！',
-            onRemovePlayer: '{name}の恐怖が和らいだ',
-            onRemoveBoss: '{name}の恐怖が和らいだ'
-        }
-    },
-    {
-        type: StatusEffectType.Oblivion,
-        name: '忘却',
-        description: 'スキルの使用方法を忘れ、使用できない',
-        duration: 3,
-        category: 'debuff',
-        isDebuff: true,
-        modifiers: {
-            canUseSkills: false
-        },
-        messages: {
-            onApplyPlayer: '{name}はスキルの使用方法を忘れてしまった！',
-            onApplyBoss: '{name}はスキルの使用方法を忘れてしまった！',
-            onRemovePlayer: '{name}はスキルの使用方法を思い出した',
-            onRemoveBoss: '{name}はスキルの使用方法を思い出した'
-        }
-    },
-
-    // Underground Worm status effects
-    {
-        type: StatusEffectType.Petrified,
-        name: '石化',
-        description: '石のように固まって行動できない',
-        duration: 2,
-        category: 'debuff',
-        isDebuff: true,
-        modifiers: {
-            canAct: false
-        },
-        messages: {
-            onApplyPlayer: '{name}は石のように固まってしまった！',
-            onApplyBoss: '{name}は石のように固まってしまった！',
-            onRemovePlayer: '{name}の石化が解けた',
-            onRemoveBoss: '{name}の石化が解けた'
-        }
     }
 ];
